+++ crates/gen/src/parser/mod.rs
@@ -19,6 +19,7 @@ mod traits;
 mod type_flags;
 mod type_kind;
 mod type_reader;
+mod type_tree;
 
 pub use blob::*;
 pub use codes::*;
@@ -39,3 +40,4 @@ pub use traits::*;
 pub use type_flags::*;
 pub use type_kind::*;
 pub use type_reader::*;
+pub use type_tree::*;
